+++ source/core/rendering/integrator/photon/photon_importance.hpp
@@ -17,8 +17,8 @@ class Importance {
     void export_heatmap(std::string_view name) const noexcept;
 
   private:
-    int2   dimensions_;
-    float* importance_;
+    int2      dimensions_;
+    uint32_t* importance_;
 
     int2   dimensions_back_;
     float2 dimensions_float_;
